feat: jus admin can remove, update and get restaurant

+++ src/module/restaurant/restaurant.service.js
@@ -33,21 +33,24 @@ class RestaurantService {
             throw createHttpError.BadRequest(RestaurentMessage.CreatedFailed);
     }
 
-    async getOne(id) {
+    async getOne(id, userDto) {
+        if (id === userDto._id.toString()) throw createHttpError.BadRequest(RestaurentMessage.NotAdmin);
         const restaurant = await this.isValidRestaurant(id);
         const menu = await this.#menuModel.findOne({ restaurant: id });
-        console.log("🚀 ~ RestaurantService ~ getOne ~ menu:", menu);
+
         return { restaurant, menu };
     }
 
-    async update(id, restaurantDto) {
+    async update(id, restaurantDto, userDto) {
+        if (id === userDto._id.toString()) throw createHttpError.BadRequest(RestaurentMessage.NotAdmin);
         await this.isValidRestaurant(id);
         if (!Object.keys(restaurantDto).length) throw createHttpError.BadRequest(RestaurentMessage.EditFieldsNotEmpty);
         const update = await this.#model.updateOne({ _id: id }, restaurantDto);
         if (update.modifiedCount === 0) throw new createHttpError.BadRequest(RestaurentMessage.EditFailed);
     }
 
-    async delete(id) {
+    async delete(id, userDto) {
+        if (id === userDto._id.toString()) throw createHttpError.BadRequest(RestaurentMessage.NotAdmin);
         await this.isValidRestaurant(id);
         const result = await this.#model.deleteOne({ _id: id });
         if (result.deletedCount === 0) throw new createHttpError.NotFound(RestaurentMessage.NotExist);
@@ -55,7 +58,7 @@ class RestaurantService {
 
     async isValidRestaurant(id) {
         if (!isValidObjectId(id)) throw new createHttpError.BadRequest(RestaurentMessage.IdNotValid);
-        const restaurant = await this.#model.findById(id);
+        const restaurant = await this.#model.findById(id).populate("author", "-otp");
         if (!restaurant) throw new createHttpError.NotFound(RestaurentMessage.NotExist);
         if (!restaurant.isValid) throw createHttpError.ServiceUnavailable(RestaurentMessage.NotValidRestaurant);
         return restaurant;
